♻️ refactor: chatHandler 일부 분리
