+++ src/mocks/chat/index.ts
@@ -1,19 +1,23 @@
 import {
   getChatsHandler,
   deleteOurChatHandler,
-  getNoticesHandler,
-  postNoticeHandler,
-  postNoticeFoldHandler,
-  postNoticeInvisibleHandler,
-  getArchivedChatHandler,
-  postArchivedChatHandler,
-  deleteArchivedChatHandler,
   getPhotosHandler,
   getPhotoDetailHandler,
   postPutGalleryHandler,
   getUploadUrlHandler,
   createPhotoHandler,
 } from './chatHandler';
+import {
+  getArchivedChatHandler,
+  postArchivedChatHandler,
+  deleteArchivedChatHandler,
+} from './chatArcivedHandler';
+import {
+  getNoticesHandler,
+  postNoticeHandler,
+  postNoticeFoldHandler,
+  postNoticeInvisibleHandler,
+} from './chatNoticeHandler';
 import {
   getQuestionsHandler,
   postQuestionsHandler,
